Connection: provide debug-message when CM crashes
Previously, when a TpConnection was invalidated due to a crash,
tp_connection_get_detailed_error() would return DBUS_ERROR_NO_REPLY
(which is plausible) but would not include any details in the dict o'
details. This is a shame, because the invalidation GError in this case
contains a message like “Name owner lost (service crashed?)”.

So, let's pass the message on.

Related-to: <https://bugs.freedesktop.org/show_bug.cgi?id=40129>
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>