codegen: add a whitelist of tp_cli_*_run_* methods
We have to generate a bunch of (deprecated) tp_cli_*_run_* methods, for
backwards-compatibility. But there's no reason to add any *more* every
time we define a new channel interface.

So here, we generate a list of methods we need to generate for backwards
compatibility (based on their being listed in the documentation), and
modify the code generator to refuse to generate any _run_ method not
named in that file.