TpChannel: don't rely on introspection queue to add the iface ID of the type
tp-glib uses to rely on its introspection queue to add the interface ID
of its channel type even when the type was already known during
construction (which is basically alway the case now as we always pass the
immutable properties when creating a TpChannel).

This was forcing TpChannel subclasses to have a CORE feature to connect
signals on their channel type interface for no good reason.

https://bugs.freedesktop.org/show_bug.cgi?id=41729