Prevent circular introspection dependency
As part of preparing a Connection the self contact will be retrieved. If
the contacts requested features include _CAPABILITIES _and_ the
connection doesn't support ContactCapabilities then the Capabilties of
the connection get use. But if we wait for this feature to be prepared
then the introspection stalls.

So instead use internal API to force getting the capabilities of the
connection regardless of the state of the various features and break the
circular dependency.