ft-channel: ensure incoming connections are actually closed (and not leaked)
Previously, we were getting the GSocketConnection and using it, but
once the splice operation was complete, we were only closing one side
of it, and then leaking the object in the process.

This is a problem because CMs can wait for EOF on the stream (like
salut) so if the stream isn't closed properly transfers can hang. It
wasn't noticed properly before because other CMs automatically close
the connection when the number of received bytes looks sane (like
gabble).

And yes, also we were leaking the GSocketConnection object.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>