channel-contacts.c: Fix crash when preparing TpCallChannel
If if a new contact fetch is queued from the callback of last
contact fetch, it leads to a crash. This is because when queueing
the new item it is processed right away since it is the only item
in queue, then when it returns from g_simple_async_result_complete()
it calls process_contacts_queue() again so the item is prepared twice,
and then freed twice.

Fix this by keeping currently being processes item outside the queue.