base-channel: deprecate interfaces class property for get_interfaces() method
With the interfaces class property, programmers writing a concrete class
could not chain-up to inherit the interfaces implemented by the base
class. Using a get_interfaces() class method, the programmer can simply
add the interfaces she is implementing to those of the base class.

Furthermore, classes that only implement an interface in specific
instances required a subclass that had to be instantiated by the
manager. This is no longer required.

For backwards compatibility, the base get_interfaces() method returns
the value of the interfaces class property.