Complete results of TpProxy D-Bus calls in an idle
GAsyncResult guarantees to call your callback from the main loop.
For historical reasons (basically "5 years ago I didn't know any
better"), TpProxy does not: if the interface is missing or the proxy
has been invalidated, the callback is called re-entrantly. I'm going
to fix that in Telepathy 1.0, but for now, let's give GAsyncResult
the correct semantics.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=45514
Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>