TpAccount: ensure that async-returned objects live as long as the result
It is valid to keep a GAsyncResult for as long as you like, so it will
have to take a copy of the result data. Otherwise, a change as simple as
replacing g_simple_async_result_complete with ..._complete_in_idle
will result in the data having already been freed by the time the
caller sees it.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=45554
Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>