Avoid forward-declaring Call classes
Having more than one typedef for the same name is an error.

If forward declarations are needed to break cycles, the typedef must
be *moved* (not copied) from the header it'd normally appear in to the
header with the forward declaration: for instance, TpCallContent
(forward-declared in call-channel.h) gets this right. In extreme cases,
a global "types.h" header can forward-declare everything, although that
often leads to unnecessary compilation, so we've avoided it in
telepathy-glib.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=46384
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>