Split library into main, dbus and core
Eventually, the dbus part will contain all the generated code, core
will contain enough of TpProxy to support that, and the main part will
contain all the high-level API.

At the moment, circular dependencies mean that we have to keep almost
everything in the main part. The core part contains the error quark
for TpError and the generated enums, and the dbus part contains the
quarks for interface names.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=46835
Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>