tp_protocol_dup_params, tp_protocol_dup_param, tp_protocol_borrow_params: add
tp_protocol_dup_params might be less nice for C (you have to free the
list and the items), but is definitely nicer for Python and other g-i
bindings. It will probably be renamed to ...get_params in Telepathy 1.0.

tp_protocol_borrow_params is an efficient "C binding", but not as GObject'y.

tp_protocol_dup_param will replace get_param in Telepathy 1.0.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=46358
Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>