Redo documentation setup every time a copied source file changes
gtk-doc copies these files from the srcdir into the builddir, but after
it does so, it touches setup-build.stamp. That file has no prerequisites
in gtk-doc.mk, so once these files have been copied once, they won't be
updated from the srcdir unless you explicitly delete them. This is pretty
annoying when you're adding API on branches and building out-of-tree.

To fix this, we just need to give it some prerequisites.

Reviewed-by: Xavier Claessens <xavier.claessens@collabora.co.uk>