improve the mtime-transferring hack in tools/
The point of this stuff is to rebuild generated code when, for instance,
libtpcodegen.py has changed, without every rule that uses
glib-something-gen.py having to know to depend on libtpcodegen.py as well.
To do that, we propagate newer mtime from Python library code to Python
executables, using touch(1).

If building out-of-tree, we need to touch the file in srcdir,
rather than creating an empty file in the builddir. The rules that
depend on glib-something-gen.py specifically depend on the version in
the $(srcdir), so in an out-of-tree build, if we aren't careful to act
on the version in the srcdir, editing libtpcodegen.py won't actually
cause a rebuild.

Reviewed-by: Xavier Claessens <xavier.claessens@collabora.co.uk>