Generate reentrant-methods.list from ABI lists, not documentation
The ABI lists are really the canonical list of what we've committed
to generating, and we don't want every alteration to -sections.txt
to trigger a rebuild. It currently does, because reentrant-methods.list
causes a rebuild of the client codegen, which causes a rebuild of
everything that includes channel.h, connection.h or telepathy-glib.h.