dbus-tube test: return TRUE from ::new-connection handler
The documentation for GDBusServer::new-connection says:

> If you want to accept the connection, take a reference to the connection
> object and return TRUE.

We were not doing this—the callback returned void. So random bytes from
the stack were interpreted as a boolean. The result on my machine was
that this test case passed when run in isolation (-p /dbus-tube/offer)
but failed when test-dbus-tube ran as a whole. This is presumably
because GDBusServer does this:

    if (claimed)
      g_dbus_connection_start_message_processing (data->connection);

So if we happened to "return" a false value, the connection never got
started.