DBusTubeChannel: always use CREDENTIALS for now
Unlike with stream tubes, I don't think this is something which tp-glib
can decide for you. CREDENTIALS means “Use D-Bus's normal SASL
handshaking”; LOCALHOST means “turn off the same-user restriction and
optionally enable the ANONYMOUS mechanism if the socket really is only
available to local users” or something.

(To be honest, it's not obvious how an application can decide what it
wants without being patched. I think this functionality exists for
Sugar's benefit, where the CMs are run as different users to the
applications; the applications can only talk to them, IIRC, because the
session bus has authentication turned off.)