channel-contacts: reverse ownership of ContactsQueueItem and result
Previously, the ContactsQueueItem assumed that it owned the only
reference to the GSimpleAsyncResult. This could result in the
GSAR still existing after the CQI had been freed, if someone else took
a reference to the GSAR.

Refcounting the CQI would not solve this problem, because then the CQI
and the GSAR would have a cyclic reference.

Instead, change the queue of CQIs into a queue of GSARs, and have the
CQI owned by the GSAR. This means the GSAR's refcounting protects the
CQI from being freed prematurely.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=45554
Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>