dbus-tube-chan test: listen on a random address appropriate for the OS
Instead of hard-coding the abstract Unix socket '\000dbus-tube-test',
listen on a random (abstract or real) Unix socket in /tmp if on Unix,
or a random port on 127.0.0.1 if not. These are the same listening
addresses I used in test/loopback.c in libdbus. The Unix version also
matches what the session bus normally uses.

This fixes the test on non-Linux Unix (not insisting on
abstract-namespace sockets working), when run twice in parallel on Linux
(not insisting on using a particular hard-coded name), and in theory
also on Windows (not trying to use Unix sockets).

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=48600
Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>