Do not use _tp_connection_lookup_contact() outside contact.c
That function is dangerous because it does not guarantee that the returned TpContact
has an identifier set.

In the case a TpContact is being created with tp_connection_get_contacts_by_id(),
the first step is to inspect its handle and then create the TpContact object
*without* giving its ID. The id will be set later once we got the contact attributes.

If _tp_connection_lookup_contact() is called for the same handle/id between
the moment the TpContact is created and the moment it sets its attributes, then
a contact with no identifier is returned.