Add and test functions to check for particular file transfer properties
These functions make the simplifying assumption that all of these
properties are orthogonal, so if both (say) InitialOffset and Description
are supported, then they'll be supported in the same requestable
channel class.

This is about the best you can do without exposing the
requestable channel classes themselves as API and having applications
iterate through them, which seems lower-level than we want.

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=48780