codegen: inline type-checks into the generated header file
We don't want tp_cli_channel_foo() (in the -dbus library) to depend on
TP_IS_CHANNEL, which expands to something involving tp_channel_get_type()
(in the -main library). So, do the type-check in an inline wrapper around
the real function.

As a double-check, we should make sure that the proxy is at least a
proxy. Every generated function calls tp_proxy_borrow_interface_by_id()
early on, so we can use that to host the type-check.

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Jonny Lamb <jonny.lamb@collabora.co.uk>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=46835