log-walker: Keep a list of (iter, count) tuples
This represents the order in which events were handed over to the user
from each iterator. Lets say we have 10 events numbered from 0 to 9,
with 0 being the oldest and 9 the latest event, and we have 3
iterators Ia, Ib and Ic. If they were returned in batches of 5 as:
+-----+--------+
|event|iterator|
+-----+--------+
|  5  |   Ia   |
|  6  |   Ib   |
|  7  |   Ia   |
|  8  |   Ia   |
|  9  |   Ic   |
+ - - + - -- - +
|  0  |   Ic   |
|  1  |   Ic   |
|  2  |   Ib   |
|  3  |   Ib   |
|  4  |   Ia   |
+-----+--------+
Then the list would be:
(Ic, 2), (Ib, 2), (Ia, 2), (Ib, 1), (Ia, 2), (Ic, 1)

Fixes: https://bugs.freedesktop.org/41772