log-walker: Rework get_events
... so that only the caches are filled asynchronously and the rest of
the work does not involve the use of a separate thread. This is the
first step towards ensuring that we do not run the TplLogEventFilter
from a separate thread.

NB: This does not solve the actual problem. The TplLogEventFilter is
still invoked from a separate thread. However this refactoring lets us
move in that direction.

Fixes: https://bugs.freedesktop.org/54270