log-walker: Run the filter synchronously in the walker
This ensures that the TplLogEventFilter is always run from the same
thread which invoked the walker. This is implemented by keeping track
of skipped events in the history instead of silently ignoring them
within the LogIters. This has the nice side effect that we do not need
to run the filter while rewinding.

Fixes: https://bugs.freedesktop.org/54270