TpProtocol: Deprecate _borrow_ functions
New transfer and naming policy has been discussed in
https://bugs.freedesktop.org/show_bug.cgi?id=39189 and is
documented there: http://telepathy.freedesktop.org/wiki/Style/TelepathyGLib

There is intentionally no tp_protocol_get_params() because it is internally
an unintrospectable C-array, so a proper container needs to be created. In
next branch they are internally in a GPtrArray so a _get_ function could
be added.