TpConnection: connect to SelfHandleChanged much earlier
Previously, we only connected to it just before calling GetSelfHandle -
but on the "modern" code path we no longer call that, because
GetAll("...Connection") is just as informative. This meant we missed
self-handle changes on modern connection managers.

This was masked by the fact that the self-handle regression test
deliberately breaks the "modern" code path, because some of its tests
rely on GetSelfHandle being called.

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=55666
Reviewed-by: Xavier Claessens <xavier.claessens@collabora.co.uk>