TpConnection: don't block introspection if the self-contact changes
If the self-contact changes while we're introspecting it,
the upgrade callback for the old self-contact does nothing, and we'll
wait forever - unless we introspect the *new* self-contact, to get
back on track.

Unfortunately, we deleted the test-case for this on the 1.0 branch.
It's a little harder to test, because introspecting the self-contact
is usually instantaneous (the exception being if we force a round-trip
while adding features).