Remove pkg-config dependency from .pc files
This causes problems for Yocto, since they use pc files when choosing
package dependencies. It doesn't make a huge amount of sense for a
.pc file to depend on pkg-config: by the time you've parsed the file,
you've got pkg-config. And the versioned dependency is checked by
configure, as commit 1154794 (which added this) mentions, so if you've
compiled telepathy-glib, your pkg-config is new enough.