log-walker: Use a queue instead of a mutex for serializing operations
Earlier we were using threads to implement the asynchronous get_events
and rewind methods, so we used a mutex to serialize them. However, now
get_events has been moved to a single threaded model, so using a mutex
will lead to undefined behaviour.

Instead we use a queue to store incoming operations (ie. get_events or
rewind) and they are executed one after the other.

Fixes: https://bugs.freedesktop.org/54270