Annotate tp_account_update_parameters_finish properly.
kuuko discovered that unset_parameters is wrongly inferred to be a
string argument, not a string array argument. At runtime, passing a
string crashes. I think this is a bug in gobject-introspection but we
can annotate it as an array to fix this particular function.

(cherry picked from commit 66982e6c17691869aa17c00416d61674b5b3e34f)
Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>