with-session-bus.sh: clear DBUS_STARTER_*
telepathy-glib prefers to use DBUS_BUS_STARTER, for the benefit of
Telepathy components that are (or might be) service-activatable.
The actual terminal part of gnome-terminal 3.8 is an activatable service,
so when run from there, we pick up the real session bus' address via
DBUS_STARTER_ADDRESS. We don't want that.

(Yes, it's DBUS_STARTER_ADDRESS, not DBUS_STARTER_BUS_ADDRESS. The
reference implementation and the D-Bus Specification both say so.)

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=63119
Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Xavier Claessens <xavier.claessens@collabora.co.uk>