TpConnectionManager: retry introspection after CM exits, up to once
Many connection managers automatically exit after 5 seconds of
inactivity. If the CM has no .manager file *and* exits in this way
while we are introspecting it, we would previously consider it to have
failed introspection - but with sufficiently unfortunate timing,
that can result in empathy-accounts not considering Haze to exist.

To avoid this, without going into an infinite loop if the CM fails to
introspect, retry once, but only once.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=67183
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>