cm test: test "drop name on get" on a modern CM
We were only testing this case on the archaic CM, but that's
undesirable when converting to Telepathy 1.0, since a CM with no
D-Bus properties will cease to be a valid CM.

The diff is a little confusing because I'm basically turning
test_dbus_fallback() and its fork test_dbus_ready() back into
the same function. The merged version does everything that
either of its precursors did.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=69283
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>