TpConnectionManager: retry introspection after CM exits, up to once
Many connection managers automatically exit after 5 seconds of
inactivity. If the CM has no .manager file *and* exits in this way
while we are introspecting it, we would previously consider it to have
failed introspection - but with sufficiently unfortunate timing,
that can result in empathy-accounts not considering Haze to exist.

To avoid this, without going into an infinite loop if the CM fails to
introspect, retry once, but only once.

[This is a forward-port of commits 60120429b and 246e201a
to Telepathy 1.0.]

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=67183
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=69283
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>