Require Automake 1.12 and force use of the parallel tests driver
This means we can use the newly-recommended AM_TESTS_ENVIRONMENT
for the environment (TESTS_ENVIRONMENT is now reserved for the user,
like CFLAGS), and LOG_COMPILER for the executable that will wrap
the logs.

We also no longer need test-wrapper.sh for the tests, because
Automake has similar functionality built-in (on a buildbot or whatever,
use "make check VERBOSE=1" to cat the logs automatically).

Also use subdir-objects, to shut Automake 1.14 up.