Use the Automake "parallel-tests" driver, albeit in non-parallel mode
This seems perverse, but it means we invoke tests in the same way
I want to use for other Telepathy projects, which will help when
we want to merge their source trees together.

Now that we can rely on having the parallel-tests setup, we can use
AM_TESTS_ENVIRONMENT instead of TESTS_ENVIRONMENT (which is now
reserved for the user, like CFLAGS), and we must use LOG_COMPILER
rather than TESTS_ENVIRONMENT for "adverb" command prefixes.

We no longer need to use test-wrapper.sh for the tests, because
Automake has similar functionality built-in. On a buildbot
or similar, use "make check VERBOSE=1" to cat the logs automatically.