call-cancellation, disconnection tests: don't use TpDBusDaemon as proxy
TpDBusDaemon is rather special: in particular, it can't be invalidated
implicitly by the peer disconnecting, because there is no peer.
At the moment we fake it by forcing the DBusGProxy to be disposed,
but that's an implementation detail; it would be better for the peer
to actually disconnect.

While we're changing all these lines anyway, make a step towards
using GTest, by grouping some of the miscellaneous state in a
fixture.