call-cancellation and disconnection tests: don't dispose DBusGProxy
Using g_object_run_dispose() to force the DBusGProxy to be disposed
was really exploiting an implementation detail of dbus-glib, and
GDBus doesn't have an equivalent. To keep this test working
after porting to GDBus, we need to do it properly, by having a private
D-Bus connection and closing it.