contacts test: don't create contacts til after the signal we want to avoid
In each of these test cases we're trying to create a contact that
does not know a particular piece of information. However, TpContact
opportunistically fills in pieces of information that it sees in
signals that go past, and tp_tests_contacts_connection_change_*()
do emit signals.

Previously, we got away with it; I'm not sure how. Perhaps
tp_connection_upgrade_contacts_async() always returned before the
signal was received?