account-request test: prepare account manager
We're using a TpAccountManager that came from tp_account_manager_dup(),
which means consecutive test-cases can use the same object.
The TpAM calls GetAll in _tp_account_manager_constructed().
If we don't wait for it to be prepared, then that GetAll call might
fail, during the window in which our account manager service is
unregistered.

Under dbus-glib, I don't think that can happen, because we don't
iterate the main loop between dropping the old service in teardown
and starting the new one in setup; but under GDBus it can, because
the GDBus worker thread keeps going.