tests/dbus/disconnection: close connection before emitting signal
dbus-glib does the actual I/O in the main thread, but we defer signal
processing to an idle. The close is not deferred in this way, so the
DBusGProxy will be marked as destroyed before the idle processes the
signal, and the signal is not received.

GDBus does all of its actual I/O on another thread, and defers
everything to an idle, so closing the connection just appends a "closed"
event to the queue, where it appears *after* the signal, and the signal
is received before we process it.