Remove TpProxy::interface-added, and dbus-glib marshaller registration
In the glorious GDBus future, we won't need marshaller and signal
registration.

This also means we can get rid of the init_known_interfaces() family
of functions, which were only there to support interface-added.

This commit is the first of many that will break bisection.
That's somewhat unavoidable if you want to go from libdbus to GDBus
without having a single mega-commit :-)