TpProxy: rewrite signal listening
As with async calls, we now use GVariant, don't take ownership of
the arguments in the wrapper function, do more work in
hand-written code and less work in generated code, and don't
need to mess about with DBusGProxy or idles because GDBus is just
better.

I've moved monitoring the TpProxy's lifetime from "signal connection
listens for invalidated signal" to "TpProxy explicitly drops signal
connections" in order to avoid depending on dispose emitting
invalidated, which was always a design flaw really. That also gives
us the opportunity to move the invocation of "drop all signal
connections" into an idle, which is necessary so that other listeners
to the same signal that invalidated the object (e.g. StatusChanged)
will still be notified.