name owner watching: more debug, and use a free-function to free watches
Several Mission Control tests crash when trying to add more callbacks
to a _NameOwnerWatch that has not been removed from the hash table
after calling _tp_dbus_daemon_stop_watching(). It's not clear to me
how that can happen, but turning _tp_dbus_daemon_stop_watching()
into an ordinary free-function ensures that "remove it from the hash
table" and "free it" happen in the right order.