Fold the core library into the dbus library
tp_dbus_g_method_return_not_implemented is non-essential, so we can
easily put it in the -dbus library.

tp_proxy_check_interface_by_id should be available in the high-level
API, but with a couple of #define hacks we can make it available to
generated code in the -dbus library under the same name.

tp_proxy_pending_call_v1_new and tp_proxy_signal_connection_v1_new
can safely go in the -dbus library: they don't deserve to be part of
the high-level API.

This reduces us to one shared library per level of API stability,
which is exactly the right number.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=76855
Reviewed-by: Xavier Claessens