contacts/subscription-states test: redesign to fix a race condition
Under GDBus, we can receive more than one signal in the same iteration
of the main loop. I missed this instance in my initial GDBus port
because it's inconsistent and hard to reproduce, but it's the same
anti-pattern: we're waiting for two signals that both happen as a
result of the same action:

    signal_cb (...)
    {
      g_main_loop_quit (loop);
    }
    ...
    test (...)
    {
      ...
      g_main_loop_run (loop);
      ... things without side-effects ...
      g_main_loop_run (loop);
      ...
    }

I think it's better style for test code to be more explicit and have
less spooky-action-at-a-distance, anyway.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=77139
Reviewed-by: Xavier Claessens