channel-introspect: make a test more realistic
We're using an Entity_Type_None (nameless chatroom) channel, but
claiming in its immutable properties that it has a non-empty
target of type Contact. That means we're effectively making assertions
about whether the immutable properties we specified at the beginning,
or the result of GetAll(), will "win", which makes this test
unnecessarily brittle.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=77139
Reviewed-by: Xavier Claessens